+++ lecture_2/hw/shop_api/shop/cart.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel
-from hw.shop_api.shop.Items_in_cart import Items_In_Cart
+from lecture_2.hw.shop_api.shop.Items_in_cart import Items_In_Cart
 
 
 class Cart(BaseModel):
